+++ game/components/sprite_component.cpp
@@ -15,7 +15,7 @@
 using namespace augs;
 
 namespace components {
-	void sprite::drawing_input::set_global_time_seconds(const float secs) {
+	void sprite::drawing_input::set_global_time_seconds(const double secs) {
 		global_time_seconds = secs;
 	}
