+++ game/detail/basic_renderable_drawing_input.h
@@ -17,7 +17,7 @@ struct basic_renderable_drawing_input : vertex_triangle_buffer_reference {
 	rgba colorize = white;
 	renderable_drawing_type drawing_type = renderable_drawing_type::NORMAL;
 
-	void set_global_time_seconds(const float) {
+	void set_global_time_seconds(const double) {
 
 	}
 };
