+++ game/detail/gui/game_gui_context.h
@@ -113,7 +113,7 @@ public:
 		camera_cone camera,
 		const aabb_highlighter& highlighter,
 		const interpolation_system& interp,
-		const float interpolation_ratio,
+		const double interpolation_ratio,
 		const input_context& input_information,
 		augs::vertex_triangle_buffer& output
 	) :
@@ -133,7 +133,7 @@ public:
 	const aabb_highlighter& highlighter;
 	const interpolation_system& interp;
 	const input_context& input_information;
-	const float interpolation_ratio;
+	const double interpolation_ratio;
 
 	augs::vertex_triangle_buffer& get_output_buffer() const {
 		return output;
