+++ game/view/viewing_session.cpp
@@ -257,7 +257,7 @@ void viewing_session::view(
 	const cosmos& cosmos,
 	const entity_id viewed_character,
 	const visible_entities& all_visible,
-	const float interpolation_ratio,
+	const double interpolation_ratio,
 	const augs::network::client& details
 ) const {
 	using namespace augs::gui::text;
@@ -273,7 +273,7 @@ void viewing_session::view(
 	const cosmos& cosmos,
 	const entity_id viewed_character,
 	const visible_entities& all_visible,
-	const float interpolation_ratio,
+	const double interpolation_ratio,
 	const augs::gui::text::formatted_string& custom_log
 ) const {
 	frame_profiler.new_measurement();
